IskraGPRS: CSD call will check for the last update and will not fail if two updates have happened after the call. IskraP2LPC: Profiles with data in the future will be skipped, the rest of the intervals will stay. NTA protocol: Decommissioning of the mbus meter will result in a deletion of the gateway NTA protocol: Apply threshold message is changed to LoadLimit (naming issue) -> related to this, there has been an attempt to generalize the messages, GenericMessaging contains several generic(also used for classic protocols) messages, the protocol needs to use these so all naming values will be the same. NTA protocol: Mbus register groups are supported Mantis issue: 5074,5068,4969,4958,4956
git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@35280 0d108963-8a47-0410-809b-ebd9fb3b3407

+++ upl/com.energyict.mdc.upl.impl/protocolimpl/src/main/java/com/energyict/genericprotocolimpl/webrtukp/eventhandling/PowerFailureLog.java
@@ -12,6 +12,13 @@ import com.energyict.dlms.axrdencoding.OctetString;
 import com.energyict.dlms.axrdencoding.util.AXDRDateTime;
 import com.energyict.protocol.MeterEvent;
 
+/**
+ * 
+ * @author gna
+ * Changes:
+ * GNA|20072009| Changed the duration to a long, otherwise you could get negative durations ...
+ */
+
 public class PowerFailureLog {
 	
 	private TimeZone timeZone;
@@ -28,7 +35,7 @@ public class PowerFailureLog {
 		int size = this.dcEvents.getRoot().getNrOfElements();
 		Date eventTimeStamp = null;
 		for(int i = 0; i <= (size-1); i++){
-			int duration = (int)this.dcEvents.getRoot().getStructure(i).getValue(1);
+			long duration = this.dcEvents.getRoot().getStructure(i).getValue(1);
 			if(isOctetString(this.dcEvents.getRoot().getStructure(i).getElement(0))){
 				eventTimeStamp = new AXDRDateTime(new OctetString(dcEvents.getRoot().getStructure(i).getOctetString(0).getArray())).getValue().getTime();
 			}
@@ -39,7 +46,7 @@ public class PowerFailureLog {
 		return meterEvents;
 	}
 
-	private void buildMeterEvent(List<MeterEvent> meterEvents, Date eventTimeStamp, int duration) {
+	private void buildMeterEvent(List<MeterEvent> meterEvents, Date eventTimeStamp, long duration) {
 		meterEvents.add(new MeterEvent(eventTimeStamp, MeterEvent.PHASE_FAILURE, "Duration of power failure: " + duration));
 	}
